Turn off fail fast
When fail fast is on, any errors that occur when deploying a flow result in all other flows being deployed cancelling.

When a second commit is pushed fixing the previous error, the cancelled flows aren't deployed as they no longer have “new” code in the recent commit. 

By setting `fail-fast: false` it will stop the auto cancelling behaviour and allow flows to be deployed irrespective of another failing.